requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==5.1.5
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
multidict==6.1.0
mysqlclient==2.2.7
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
propcache==0.2.1
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
stripe==11.6.0
twilio==9.4.5
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.1
whitenoise==6.9.0
yarl==1.18.3
